requirements.txt
absl-py==2.0.0
anyio==3.6.2
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
arrow==1.2.3
asgiref==3.6.0
asttokens==2.2.1
astunparse==1.6.3
attrs==22.2.0
backcall==0.2.0
beautifulsoup4==4.11.1
bleach==5.0.1
cachetools==5.3.2
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
ChatterBot==1.0.4
chatterbot-corpus==1.2.0
cleaner==0.4.0
click==8.1.3
clize==5.0.0
colorama==0.4.6
comm==0.1.2
contourpy==1.0.7
coverage==5.5
cycler==0.11.0
debugpy==1.6.5
decorator==5.1.1
defusedxml==0.7.1
deprecation==2.1.0
distro==1.8.0
Django==4.1.7
dm-tree==0.1.8
docopt==0.6.2
docutils==0.19
entrypoints==0.4
et-xmlfile==1.1.0
executing==1.2.0
eyed3==0.9.7
fastjsonschema==2.16.2
filelock==3.12.2
filetype==1.2.0
Flask==2.2.2
Flask-MySQLdb==2.0.0
flatbuffers==23.5.26
fonttools==4.38.0
fqdn==1.5.1
fsspec==2023.12.2
gast==0.5.4
google==3.0.0
google-auth==2.26.1
google-auth-oauthlib==1.2.0
google-pasta==0.2.0
graphviz==0.20.1
grpcio==1.60.0
h5py==3.10.0
idna==3.4
ipykernel==6.20.2
ipython==8.8.0
ipython-genutils==0.2.0
ipywidgets==8.0.4
isoduration==20.11.0
itsdangerous==2.1.2
jedi==0.18.2
Jinja2==3.1.2
joblib==1.2.0
jsonpointer==2.3
jsonschema==4.17.3
jupyter==1.0.0
jupyter-console==6.4.4
jupyter-events==0.6.3
jupyter_client==7.4.9
jupyter_core==5.1.3
jupyter_server==2.1.0
jupyter_server_terminals==0.4.4
jupyterlab-pygments==0.2.2
jupyterlab-widgets==3.0.5
keras==2.15.0
kiwisolver==1.4.4
libclang==16.0.6
Markdown==3.5.1
markdown-it-py==3.0.0
MarkupSafe==2.1.1
mathparse==0.1.2
matplotlib==3.6.3
matplotlib-inline==0.1.6
mdurl==0.1.2
mistune==2.0.4
ml-dtypes==0.2.0
MouseInfo==0.1.3
mpmath==1.3.0
mysqlclient==2.2.4
namex==0.0.7
nbclassic==0.4.8
nbclient==0.7.2
nbconvert==7.2.8
nbformat==5.7.3
nest-asyncio==1.5.6
networkx==3.1
newsapi==0.1.1
nltk==3.8
nodejs==0.1.1
notebook==6.5.2
notebook_shim==0.2.2
npm==0.1.1
numpy==1.26.3
oauthlib==3.2.2
od==2.0.2
openpyxl==3.0.10
opt-einsum==3.3.0
optional-django==0.1.0
packaging==23.0
pandas==1.5.1
pandocfilters==1.5.0
parso==0.8.3
patsy==0.5.3
pgmpy==0.1.23
pickleshare==0.7.5
Pillow==9.3.0
Pint==0.20.1
pipreqs==0.4.11
platformdirs==2.6.2
plotly==5.14.1
prometheus-client==0.15.0
prompt-toolkit==3.0.36
protobuf==4.23.4
psutil==5.9.4
pure-eval==0.2.2
pyasn1==0.5.1
pyasn1-modules==0.3.0
pyAudioAnalysis==0.3.14
PyAutoGUI==0.9.53
pycparser==2.21
pydicom==2.3.1
pydub==0.25.1
PyGetWindow==0.0.9
Pygments==2.14.0
pymongo==3.13.0
PyMsgBox==1.0.9
pyparsing==3.0.9
pyperclip==1.8.2
PyRect==0.2.0
pyrsistent==0.19.3
PyScreeze==0.1.28
python-dateutil==2.8.2
python-json-logger==2.0.4
pytweening==1.0.4
pytz==2022.6
pywhatkit==5.4
pywin32==305
pywinpty==2.0.10
PyYAML==6.0
pyzmq==25.0.0
qtconsole==5.4.0
QtPy==2.3.0
regex==2022.10.31
requests==2.28.1
requests-oauthlib==1.3.1
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.7.0
rsa==4.9
scikit-learn==1.3.2
scipy==1.10.0
seaborn==0.12.2
Send2Trash==1.8.0
sigtools==4.0.1
six==1.16.0
sklearn==0.0.post1
sniffio==1.3.0
sounddevice==0.4.6
soundfile==0.12.1
soupsieve==2.3.2.post1
SQLAlchemy==1.2.19
sqlparse==0.4.3
stack-data==0.6.2
statsmodels==0.13.5
sympy==1.12
tabula-py==2.7.0
tenacity==8.2.2
tensorboard==2.15.1
tensorboard-data-server==0.7.2
tensorflow==2.15.0
tensorflow-estimator==2.15.0
tensorflow-intel==2.15.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.4.0
terminado==0.17.1
textblob==0.17.1
threadpoolctl==3.1.0
tinycss2==1.2.1
toml==0.10.2
torch==2.1.2
tornado==6.3.1
tqdm==4.64.1
traitlets==5.8.1
typing_extensions==4.7.1
tzdata==2022.7
uri-template==1.2.0
urllib3==1.26.12
wcwidth==0.2.6
webcolors==1.12
webencodings==0.5.1
websocket-client==1.4.2
Werkzeug==2.2.2
widgetsnbextension==4.0.5
wikipedia==1.4.0
wrapt==1.14.1
xgboost==2.0.3
xlrd==2.0.1
yarg==0.1.9
yellowbrick==1.5
